implemented the rlationship between a user and a listing

+++ app/Http/Controllers/ListingController.php
@@ -32,7 +32,7 @@ class ListingController extends Controller
 
     public function store(Request $request)
     {
-        $formFIelds = $request->validate([
+        $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
             'location' => 'required',
@@ -45,10 +45,12 @@ class ListingController extends Controller
 
         ]);
         if ($request->hasFile('logo')) {
-            $formFIelds['logo'] = $request->file('logo')->store('logos', 'public');
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        Listing::create($formFIelds);
+        $formFields['user_id'] = auth()->id();
+
+        Listing::create($formFields);
 
         return redirect("/")->with('message', 'Job Listing added successfuly!');
     }
@@ -60,6 +62,11 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+        if ($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized action');
+        }
+
+
         $formFIelds = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -83,7 +90,19 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized action');
+        }
         $listing->delete();
         return redirect("/")->with('message', 'Job Listing deleted successfuly!');
+    }
+
+    public function manage()
+    {
+
+
+
+        return view('listings.manage', ['listings' => auth()
+            ->user()->listings()->get()]);
     }
 }
