+++ routes/web.php
@@ -1,8 +1,9 @@
 <?php
 
-use App\Http\Controllers\ListingController;
-use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,9 +17,15 @@ use App\Models\Listing;
 */
 
 Route::get('/', [ListingController::class, 'index']);
-Route::post('/listings', [ListingController::class, 'store']);
-Route::get('/listings/create', [ListingController::class, 'create']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+Route::post('/users', [UserController::class, 'store']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
